+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added version to config
 - Added BuildStatus component
 - Added more font types to styleConstants
+- Added Android back handler
 
 ### Removed